+++ lib/src/controllers/flukki_controller.dart
@@ -1,7 +1,9 @@
 import 'dart:async';
 
-import '../helpers/app_version_controller.dart';
-import '../helpers/device_id_controller.dart';
+import 'package:flutter/foundation.dart';
+import 'package:flutter/material.dart';
+
+import '../helpers/user_controller.dart';
 import '../models/product_tour_model.dart';
 import 'context_controller.dart';
 import 'product_tours_controller.dart';
@@ -11,9 +13,22 @@ import 'test_product_tour_controller.dart';
 class FlukkiController {
   static final FlukkiController instance = FlukkiController._();
 
-  FlukkiController._();
+  FlukkiController._() {
+    assert(const bool.fromEnvironment('flutter.memory_allocations') == true,
+        'Environment variable not set. Add <flutter.memory_allocations=true> environment variable. Your run method should look like this "--dart-define=flutter.memory_allocations=true"');
+    MemoryAllocations.instance.addListener((event) {
+      if (event.object is Element) {
+        if (event is ObjectCreated) {
+          ContextController.instance.addElement(event.object as Element);
+        } else if (event is ObjectDisposed) {
+          ContextController.instance.removeElement(event.object as Element);
+        }
+      }
+    });
+    WidgetsBinding.instance.addPersistentFrameCallback(
+        ContextController.instance.performCheckIfPossible);
+  }
 
-  bool _wasWidgetInspectorInitiated = false;
   bool _isInBuilderMode = false;
   bool _isInBuilderTestMode = false;
   Map<String, void Function()> callbacks = {};
@@ -29,8 +44,6 @@ class FlukkiController {
 
   bool get isInBuilderTestMode => _isInBuilderTestMode;
 
-  bool get wasWidgetInspectorInitiated => _wasWidgetInspectorInitiated;
-
   Future<void> initialize(
       {required String key,
       required String appName,
@@ -38,16 +51,7 @@ class FlukkiController {
     apiKey = key;
     appId = appName;
     this.callbacks = callbacks ?? {};
-    await DeviceIdController.instance.init(key);
-    await AppVersionController.instance.init();
-    await ProductToursController.instance
-        .initialize(this.callbacks, DeviceIdController.instance.deviceId);
-    ContextController.instance.performCheck();
-    StatisticsController.instance.sendStatistics();
-  }
-
-  void widgetExplorerInitiated() {
-    _wasWidgetInspectorInitiated = true;
+    await UserController.instance.init(key);
   }
 
   TestProductTourController? testProductTourController;
@@ -55,6 +59,7 @@ class FlukkiController {
   Future<void> turnOnTestMode(ProductTour productTour) async {
     testProductTourController = TestProductTourController(productTour);
     _isInBuilderTestMode = true;
+    ContextController.instance.flushAwaiting();
     ContextController.instance.performCheck();
   }
 
@@ -71,5 +76,22 @@ class FlukkiController {
   void turnOnBuilder() {
     _isInBuilderMode = true;
     _onCreationModeChanged.add(_isInBuilderMode);
+  }
+
+  Future<void> signIn({String? userID}) async {
+    assert(apiKey != null,
+        'Plugin not initialized. Initialize plugin first, and then sign in user');
+
+    await UserController.instance.signIn(apiKey!, userId: userID);
+    await ProductToursController.instance
+        .initialize(callbacks, UserController.instance.userID);
+    await StatisticsController.instance.fetchStatistics();
+    ContextController.instance.flushAwaiting();
+    ContextController.instance.performCheck();
+  }
+
+  Future<void> signOut() async {
+    await UserController.instance.signOut();
+    await StatisticsController.instance.clear();
   }
 }