+++ lib/src/controllers/test_product_tour_controller.dart
@@ -1,11 +1,15 @@
 import '../models/product_tour_model.dart';
 
 class TestProductTourController {
-  int currentIndex = 0;
   ProductTour productTour;
 
   TestProductTourController(this.productTour) {
-    productTour.currentIndex = 0;
-    productTour.skippedIndex = null;
+    TestStatsController.currentIndex = 0;
+    TestStatsController.skippedIndex = null;
   }
+}
+
+class TestStatsController {
+  static int currentIndex = 0;
+  static int? skippedIndex;
 }