+++ CHANGELOG.md
@@ -1,3 +1,20 @@
+## 2.0.3
+
+* Updated README.md
+
+## 2.0.2
+
+* Fixed announcements preview in test mode
+
+## 2.0.1
+
+* Fixed going back to list of product tours without saving changes
+
+## 2.0.0+1
+BREAKING: Changed way of plugin initialization. User must be signed in to see product tours.
+* `signInAnonymous` and `signInUser` allows to create new sessions for users
+* `signOut` allows to sign out user
+
 ## 1.1.1
 
 * Improved widget identification